+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - concord-server, concord-console: initial support for process-,
 organization- and project-level metadata.
 
+### Changed
+
+- concord-server: apply RBAC filters when listing secrets.
+
 
 
 ## [0.86.0] - 2018-09-05